+++ src/effects/spells/magicWeapon.js
@@ -1,11 +1,44 @@
-import { baseSpellEffect } from "../specialSpells.js";
-import DDBMacros from "../DDBMacros.js";
+import { baseEnchantmentEffect } from "../effects.js";
 
-export async function magicWeaponEffect(document) {
-  let effect = baseSpellEffect(document, document.name);
-  await DDBMacros.setItemMacroFlag(document, "spell", "magicWeapon.js");
-  effect.changes.push(DDBMacros.generateMacroChange({ macroValues: "@item.level", macroType: "spell", macroName: "magicWeapon.js", priority: 0 }));
-  document.effects.push(effect);
+export function magicWeaponEffect(document) {
+  const enchantments = [
+    { bonus: "1", min: null, max: 3 },
+    { bonus: "2", min: 4, max: 5 },
+    { bonus: "3", min: 6, max: null },
+  ];
+  document.system.actionType = "ench";
+  foundry.utils.setProperty(document, "system.enchantment", {
+    restrictions: {
+      allowMagical: false,
+      type: "weapon",
+    },
+  });
+  for (const e of enchantments) {
+    let effect = baseEnchantmentEffect(document, `${document.name}: +${e.bonus}`);
+    foundry.utils.setProperty(effect, "flags.dnd5e.enchantment.level", { min: e.min, max: e.max });
+    effect.changes.push(
+      {
+        key: "name",
+        mode: CONST.ACTIVE_EFFECT_MODES.OVERRIDE,
+        value: `{}, +${e.bonus}`,
+        priority: 20,
+      },
+      {
+        key: "system.properties",
+        mode: CONST.ACTIVE_EFFECT_MODES.ADD,
+        value: "mgc",
+        priority: 20,
+      },
+      {
+        key: "system.magicalBonus",
+        mode: CONST.ACTIVE_EFFECT_MODES.OVERRIDE,
+        value: e.bonus,
+        priority: 20,
+      },
+    );
+    e.description = `This weapon has become a +${e.bonus} magic weapon, granting a bonus to attack and damage rolls.`;
+    document.effects.push(effect);
+  }
 
   return document;
 }